% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{chemical_similarity_threshold}
\alias{chemical_similarity_threshold}
\title{Find chemical similarity matches up to a threshold.}
\usage{
chemical_similarity_threshold(
  query_inchis,
  target_inchis = NULL,
  fingerprint_type = c("morgan", "topological"),
  fingerprint_args = NULL,
  threshold = 0.7
)
}
\arguments{
\item{query_inchis}{A character vector of compound inchis. Can optionally be named.
All similarities between the query and the target compounds are calculated.}

\item{target_inchis}{An optional character vector of compound inchis. If not given,
the similarities of all pairwise combinations between query compounds are calculated.}

\item{fingerprint_type}{Calculate morgan or topological fingerprints.}

\item{fingerprint_args}{Optional list of additional arguments to the RDKit fingerprinting
function.}

\item{threshold}{A double between 0 and 1 representing the minimum reported similarity.}
}
\value{
A tibble with three columns, containing names of query and target compounds
and their similarity.
}
\description{
Requires chemfp installed on the python server.
}
\examples{
chemical_similarity_threshold(
  c("resveratrol" = "InChI=1S/C14H12O3/c15-12-5-3-10(4-6-12)1-2-11-7-13(16)9-14(17)8-11/h1-9,15-17H/b2-1+"),
  c(
    "tofacitnib" = "InChI=1S/C16H20N6O/c1-11-5-8-22(14(23)3-6-17)9-13(11)21(2)16-12-4-7-18-15(12)19-10-20-16/h4,7,10-11,13H,3,5,8-9H2,1-2H3,(H,18,19,20)/t11-,13+/m1/s1",
    "aspirin" = "InChI=1S/C9H8O4/c1-6(10)13-8-5-3-2-4-7(8)9(11)12/h2-5H,1H3,(H,11,12)"
  ),
  threshold = 0.1
)
}
