% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk_df}
\alias{chunk_df}
\title{Split dataframe into chunks}
\usage{
chunk_df(df, n, seed = NULL)
}
\arguments{
\item{df}{A dataframe to be split into equal sized chunks.}

\item{n}{An integer with the desired number of chunks.}

\item{seed}{Optionally, a random seed to make chunking reproducible}
}
\value{
A list of \code{n} entries, each containing a dataframe chunk.
}
\description{
The order of entries are randomized.
}
