% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounds.R
\name{compounds}
\alias{compounds}
\title{Compound input}
\usage{
compounds(
  x,
  fingerprints = NULL,
  descriptor = c("inchi", "smiles", "smarts", "fingerprint"),
  keep_invalid = FALSE
)
}
\arguments{
\item{x}{A character vector of compounds encoded in the given way. Optionally named}

\item{fingerprints}{Optionally, a character vector of hexadecimal encoded fingerprints}

\item{descriptor}{Chemical descriptor used for encoding the compounds}

\item{keep_invalid}{Keep invalid compounds}
}
\value{
A character vector with the appropriate descriptor class set. Can be
used as input for other lspcheminf functions.
}
\description{
Compounds can be described using character vectors of SMILES, InChI or SMARTS descriptors.
For some functions fingerprint input is also accepted.
}
