% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{convert_compound_descriptor}
\alias{convert_compound_descriptor}
\title{Convert compound descriptors}
\usage{
convert_compound_descriptor(
  x,
  target_descriptor = c("smiles", "inchi", "smarts", "inchi_key"),
  url = "http://127.0.0.1:8000"
)
}
\arguments{
\item{x}{A character vector of compounds, optionally named.
Assumed to be InChI. Specify descriptor explicitly or attach
pre-computed fingerprints using \code{\link[=compounds]{compounds()}}.}

\item{target_descriptor}{The chemical descriptor targeted for conversion.}

\item{url}{URL to the lspcheminf python server}
}
\value{
A tibble with three columns, containing query compound names, query
compound descriptor and target descriptor
}
\description{
Convert compound descriptors
}
\examples{
convert_compound_descriptor(
  c(
    "tofacitnib" = "InChI=1S/C16H20N6O/c1-11-5-8-22(14(23)3-6-17)9-13(11)21(2)16-12-4-7-18-15(12)19-10-20-16/h4,7,10-11,13H,3,5,8-9H2,1-2H3,(H,18,19,20)/t11-,13+/m1/s1",
    "aspirin" = "InChI=1S/C9H8O4/c1-6(10)13-8-5-3-2-4-7(8)9(11)12/h2-5H,1H3,(H,11,12)"
  ),
  target_descriptor = "smiles"
)
}
