% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{convert_compound_identifier}
\alias{convert_compound_identifier}
\title{Convert compound identifiers}
\usage{
convert_compound_identifier(query, identifier = c("inchi", "smiles",
  "smarts"), target_identifier = c("smiles", "inchi", "smarts",
  "inchi_key"))
}
\arguments{
\item{query}{A character vector of compounds. Can optionally be named.}

\item{identifier}{The chemical identifier used for the query.}

\item{target_identifier}{The chemical identifer targeted for conversion.}
}
\value{
A tibble with three columns, containing query compound names, query
compound identifier and target identifier.
}
\description{
Convert compound identifiers
}
\examples{
convert_compound_identifier(
  c(
    "tofacitnib" = "InChI=1S/C16H20N6O/c1-11-5-8-22(14(23)3-6-17)9-13(11)21(2)16-12-4-7-18-15(12)19-10-20-16/h4,7,10-11,13H,3,5,8-9H2,1-2H3,(H,18,19,20)/t11-,13+/m1/s1",
    "aspirin" = "InChI=1S/C9H8O4/c1-6(10)13-8-5-3-2-4-7(8)9(11)12/h2-5H,1H3,(H,11,12)"
  ),
  identifier = "inchi",
  target_identifier = "smiles"
)
}
