% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{maximum_common_substructure}
\alias{maximum_common_substructure}
\title{Calculate maximum common substructure (MCS) of compounds.}
\usage{
maximum_common_substructure(x, url = "http://127.0.0.1:8000")
}
\arguments{
\item{x}{A character vector of compounds, optionally named.
Assumed to be InChI. Specify descriptor explicitly or attach
pre-computed fingerprints using \code{\link[=compounds]{compounds()}}.}

\item{url}{URL to the lspcheminf python server}
}
\value{
A character containing the common substructure of the query compounds.
}
\description{
Calculate maximum common substructure (MCS) of compounds.
}
\examples{
maximum_common_substructure(
  compounds(
    c(
      "O=C(NCc1cc(OC)c(O)cc1)CCCC/C=C/C(C)C",
      "CC(C)CCCCCC(=O)NCC1=CC(=C(C=C1)O)OC",
      "c1(C=O)cc(OC)c(O)cc1"
    ),
    descriptor = "smiles"
  )
)
}
