% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{molecular_mass}
\alias{molecular_mass}
\title{Calculate molecular mass of compounds}
\usage{
molecular_mass(compounds, url = "http://127.0.0.1:8000")
}
\arguments{
\item{compounds}{A character vector of compounds as InChI strings. Can optionally be named.}

\item{url}{URL to the lspcheminf python server}
}
\value{
A tibble with two columns, containing compound identifier and molecular weight.
}
\description{
Calculate molecular mass of compounds
}
\examples{
molecular_mass(
  c(
    "tofacitnib" = "InChI=1S/C16H20N6O/c1-11-5-8-22(14(23)3-6-17)9-13(11)21(2)16-12-4-7-18-15(12)19-10-20-16/h4,7,10-11,13H,3,5,8-9H2,1-2H3,(H,18,19,20)/t11-,13+/m1/s1",
    "aspirin" = "InChI=1S/C9H8O4/c1-6(10)13-8-5-3-2-4-7(8)9(11)12/h2-5H,1H3,(H,11,12)"
  )
)
}
