% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{murcko_scaffold}
\alias{murcko_scaffold}
\title{Calculate Murcko Scaffold of compounds.}
\usage{
murcko_scaffold(x, url = "http://127.0.0.1:8000")
}
\arguments{
\item{x}{A character vector of compounds, optionally named.
Assumed to be InChI. Specify descriptor explicitly or attach
pre-computed fingerprints using \code{\link[=compounds]{compounds()}}.}

\item{url}{URL to the lspcheminf python server}
}
\value{
A tibble with two columns, containing compound name and murcko scaffold
}
\description{
Calculate Murcko Scaffold of compounds.
}
\examples{
murcko_scaffold(
  compounds(
    c(
      "O=C(NCc1cc(OC)c(O)cc1)CCCC/C=C/C(C)C",
      "CC1=C(C=C(C=C1)NC(=O)C2=CC=C(C=C2)CN3CCN(CC3)C)NC4=NC=CC(=N4)C5=CN=CC=C5"
    ),
    descriptor = "smiles"
  )
)
}
