% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{is_organic}
\alias{is_organic}
\title{Check if compound is organic}
\usage{
is_organic(compounds, url = "http://127.0.0.1:8000")
}
\arguments{
\item{compounds}{A character vector of compounds as InChI strings. Can optionally be named.}

\item{url}{URL to the lspcheminf python server}
}
\value{
A tibble with two columns, containing compound identifier and a
boolean flag if they are organic.
}
\description{
This is true if it contains at least one carbon.
}
\examples{
is_organic(
  c(
    "tofacitnib" = "InChI=1S/C16H20N6O/c1-11-5-8-22(14(23)3-6-17)9-13(11)21(2)16-12-4-7-18-15(12)19-10-20-16/h4,7,10-11,13H,3,5,8-9H2,1-2H3,(H,18,19,20)/t11-,13+/m1/s1",
    "nitrous oxide" = "InChI=1S/N2O/c1-2-3"
  )
)
}
